Move JSON and TAP loggers to separate files
