+++ src/Codeception/Configuration.php
@@ -82,8 +82,8 @@ class Configuration
             'xml'         => 'Codeception\PHPUnit\Log\JUnit',
             'html'        => 'Codeception\PHPUnit\ResultPrinter\HTML',
             'report'      => 'Codeception\PHPUnit\ResultPrinter\Report',
-            'tap'         => 'PHPUnit\Util\Log\TAP',
-            'json'        => 'PHPUnit\Util\Log\JSON',
+            'tap'         => 'Codeception\PHPUnit\Util\Log\TAP',
+            'json'        => 'Codeception\PHPUnit\Util\Log\JSON',
             'phpunit-xml' => 'Codeception\PHPUnit\Log\PhpUnit',
         ],
         'groups'     => [],
